requirements.txt
git+https://github.com/Rapptz/discord-ext-menus.git
git+https://github.com/Gorialis/jishaku.git
git+https://github.com/XuaTheGrate/python-cse.git
git+https://github.com/ShineyDev/prettify.py.git
asyncpg==0.21.0
discord.py==1.7.1
python-decouple==3.4
python-Levenshtein==0.12.0
humanize==3.2.0
pytz==2020.5
async-cleverbot==0.2.2
psutil==5.8.0
asyncdagpi==3.2.2
wavelink==0.9.6
dateparser==1.0.0
vacefron.py==1.5.0
Pillow==8.1.1
twemoji-parser==0.4.2
pytesseract==0.3.7
CurrencyConverter==0.14.4
polaroid==0.2.10
selenium==3.141.0
Wand==0.6.5
numexpr==2.7.2
toml==0.10.2
# uvloop==0.14.0
discord-ext-ipc==1.0
ksoftapi==0.3.3a0
aiozaneapi==1.5
